Work on moving base screen logic to VEContainerScreen. Fixed BatteryBox not saving IO options
