+++ src/main/java/com/veteam/voluminousenergy/blocks/screens/VEContainerScreen.java
@@ -1,14 +1,91 @@
 package com.veteam.voluminousenergy.blocks.screens;
 
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.veteam.voluminousenergy.tools.buttons.VEIOButton;
+import com.veteam.voluminousenergy.tools.buttons.ioMenuButton;
+import com.veteam.voluminousenergy.tools.buttons.slots.SlotBoolButton;
+import com.veteam.voluminousenergy.tools.buttons.slots.SlotDirectionButton;
+import com.veteam.voluminousenergy.tools.buttons.tanks.TankBoolButton;
+import com.veteam.voluminousenergy.tools.buttons.tanks.TankDirectionButton;
+import com.veteam.voluminousenergy.tools.networking.VENetwork;
+import com.veteam.voluminousenergy.tools.networking.packets.UuidPacket;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.components.Widget;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 
+import java.util.UUID;
+
 public abstract class VEContainerScreen<T extends AbstractContainerMenu> extends AbstractContainerScreen<T> {
 
 
-    public VEContainerScreen(T p_97741_, Inventory p_97742_, Component p_97743_) {
-        super(p_97741_, p_97742_, p_97743_);
+    public VEContainerScreen(T menu, Inventory inventory, Component component) {
+        super(menu, inventory, component);
     }
+
+    public void drawIOSideHelper(PoseStack matrixStack, int i, int j, int mouseX, int mouseY, float partialTicks){
+
+        for(Widget widget : this.renderables){
+            if (widget instanceof ioMenuButton){
+                if (((ioMenuButton) widget).shouldIOBeOpen()) { // This means IO Should be open
+                    this.renderables.forEach(button ->{
+                        if (button instanceof VEIOButton){
+                            ((VEIOButton) button).toggleRender(true);
+                        }
+                    });
+                } else {
+                    this.renderables.forEach(button ->{
+                        if(button instanceof VEIOButton){
+                            ((VEIOButton) button).toggleRender(false);
+                        }
+                    });
+                }
+            }
+        }
+    }
+
+
+    public void updateButtonDirection(int direction, int slotId){
+        for(Widget widget: this.renderables){
+            if(widget instanceof SlotDirectionButton && ((SlotDirectionButton) widget).getAssociatedSlotId() == slotId ){
+                ((SlotDirectionButton) widget).setDirectionFromInt(direction);
+            }
+        }
+    }
+
+    public void updateBooleanButton(boolean status, int slotId){
+        for(Widget widget: this.renderables){
+            if(widget instanceof SlotBoolButton && ((SlotBoolButton) widget).getAssociatedSlotId() == slotId){
+                ((SlotBoolButton) widget).toggleRender(true);
+                ((SlotBoolButton) widget).setStatus(status);
+                ((SlotBoolButton) widget).toggleRender(false);
+            }
+        }
+    }
+
+    public void updateTankDirection(int direction, int id){
+        for(Widget widget: this.renderables){
+            if(widget instanceof TankDirectionButton && ((TankDirectionButton) widget).getId() == id ){
+                ((TankDirectionButton) widget).setDirectionFromInt(direction);
+            }
+        }
+    }
+
+    public void updateTankStatus(boolean status, int id){
+        for(Widget widget: this.renderables){
+            if(widget instanceof TankBoolButton && ((TankBoolButton) widget).getId() == id){
+                ((TankBoolButton) widget).toggleRender(true);
+                ((TankBoolButton) widget).setStatus(status);
+                ((TankBoolButton) widget).toggleRender(false);
+            }
+        }
+    }
+
+    public void informTileOfIOButton(boolean connection){
+        UUID uuid = Minecraft.getInstance().player.getUUID();
+        VENetwork.channel.sendToServer(new UuidPacket(uuid, connection));
+    }
+
 }
