+++ src/main/java/com/veteam/voluminousenergy/tools/networking/packets/BatteryBoxSlotPairPacket.java
@@ -1,5 +1,6 @@
 package com.veteam.voluminousenergy.tools.networking.packets;
 
+import com.veteam.voluminousenergy.VoluminousEnergy;
 import com.veteam.voluminousenergy.blocks.containers.BatteryBoxContainer;
 import com.veteam.voluminousenergy.blocks.tiles.BatteryBoxTile;
 import net.minecraft.client.Minecraft;
@@ -59,9 +60,10 @@ public class BatteryBoxSlotPairPacket {
                     if (tileEntity instanceof BatteryBoxTile) {
                         ((BatteryBoxTile) tileEntity).updateSlotPair(packet.status, packet.id);
                         tileEntity.setChanged();
+                        VoluminousEnergy.LOGGER.info(packet.status + " " + packet.id);
                     }
                 } else {
-                    ((BatteryBoxContainer) openContainer).updateSlotPairButton(packet.status, packet.id);
+                    //((BatteryBoxContainer) openContainer).updateSlotPairButton(packet.status, packet.id);
                 }
             }
         }
